Search2: Avoid TueFind inheriting from IxTheo (use trait instead)
