+++ module/IxTheo/src/IxTheo/Search/Search2/Options.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace IxTheo\Search\Search2;
+
+class Options extends \TueFind\Search\Search2\Options
+{
+
+}
