+++ module/IxTheo/src/IxTheo/Search/Search2/Params.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace IxTheo\Search\Search2;
+
+class Params extends \TueFind\Search\Search2\Params
+{
+}
