+++ module/IxTheo/src/IxTheo/Search/Search2/Results.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace IxTheo\Search\Search2;
+
+class Results extends \TueFind\Search\Search2\Results
+{
+    use \IxTheo\Search\Solr\ResultsTrait;
+}
